orderable todo list refactor
