+++ cogboard-webapp/src/components/widgets/dialogFields/ToDoListinput.js
@@ -4,18 +4,10 @@ import { v4 } from 'uuid';
 import { remove } from 'ramda';
 import { postWidgetContentUpdate } from '../../../utils/fetch';
 import { saveWidget } from '../../../actions/thunks';
-import { prepareChangeEvent } from './helpers';
-import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
-import {
-  FormControl,
-  IconButton,
-  ListItem,
-  ListItemSecondaryAction,
-  ListItemText,
-  Tooltip
-} from '@material-ui/core';
-import { Add, Check, Edit, Delete } from '@material-ui/icons';
-import { StyledFab, StyledList, StyledInput, StyledFabGroup } from './styled';
+import { prepareChangeEvent, RenderDragableList } from './helpers';
+import { FormControl } from '@material-ui/core';
+import { Add, Check, Delete } from '@material-ui/icons';
+import { StyledFab, StyledInput, StyledFabGroup } from './styled';
 
 const ToDoListInput = ({ value, values, onChange }) => {
   const [formValueItemText, setFormValueItemText] = useState('');
@@ -34,17 +26,6 @@ const ToDoListInput = ({ value, values, onChange }) => {
       };
     })
   );
-
-  const handleOnDragEnd = result => {
-    if (!result.destination) return;
-
-    const tempItems = items;
-    const [reorderedItem] = tempItems.splice(result.source.index, 1);
-    tempItems.splice(result.destination.index, 0, reorderedItem);
-
-    setItems(tempItems);
-    onChange(prepareChangeEvent(tempItems, 'array'));
-  };
 
   const resetInput = () => {
     setFormValueItemText('');
@@ -174,65 +155,15 @@ const ToDoListInput = ({ value, values, onChange }) => {
           </StyledFab>
         )}
       </StyledFabGroup>
-      <DragDropContext onDragEnd={handleOnDragEnd}>
-        <Droppable droppableId="characters">
-          {provided => (
-            <StyledList {...provided.droppableProps} ref={provided.innerRef}>
-              {items.map((item, index) => (
-                <Draggable key={item.id} draggableId={item.id} index={index}>
-                  {provided => (
-                    <ListItem
-                      ref={provided.innerRef}
-                      {...provided.draggableProps}
-                      {...provided.dragHandleProps}
-                      dense
-                      button
-                      selected={editMode === item.id}
-                      onClick={() => {
-                        handleEdit(item.id);
-                      }}
-                    >
-                      <ListItemText
-                        primary={item.itemText}
-                        style={
-                          selectedItems.includes(item.id)
-                            ? { textDecoration: 'line-through' }
-                            : {}
-                        }
-                      />
-                      <ListItemSecondaryAction>
-                        <Tooltip title="Edit" placement="bottom">
-                          <IconButton
-                            aria-label="Edit"
-                            disabled={editMode === item.id}
-                            onClick={() => {
-                              handleEdit(item.id);
-                            }}
-                          >
-                            <Edit />
-                          </IconButton>
-                        </Tooltip>
-                        <Tooltip title="Delete" placement="bottom">
-                          <IconButton
-                            aria-label="Delete"
-                            disabled={editMode === item.id}
-                            onClick={() => {
-                              handleDelete(index);
-                            }}
-                          >
-                            <Delete />
-                          </IconButton>
-                        </Tooltip>
-                      </ListItemSecondaryAction>
-                    </ListItem>
-                  )}
-                </Draggable>
-              ))}
-              {provided.placeholder}
-            </StyledList>
-          )}
-        </Droppable>
-      </DragDropContext>
+      <RenderDragableList
+        items={items}
+        setEvent={setItems}
+        onChange={onChange}
+        prepareChangeEvent={prepareChangeEvent}
+        editMode={editMode}
+        handleEdit={handleEdit}
+        handleDelete={handleDelete}
+      ></RenderDragableList>
     </FormControl>
   );
 };
